+++ src/include/game/texture.h
@@ -33,7 +33,7 @@ u32 func0f172f54(void);
 void func0f172f5c(Gfx *gdl, u32 arg1, u32 arg2);
 void func0f173010(u32 *texturenum, u32 arg1, u32 arg2);
 u32 func0f173434(void);
-void func0f1734e8(u8 arg0, void *arg1);
+void func0f1734e8(u16 arg0, void *arg1);
 u32 func0f173520(void);
 u32 func0f17353c(void);
 void texturesLoadConfigs(void);
